projects/15-justin-lee/src/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
httpx>=0.27.0
PyJWT>=2.9.0
supabase>=2.6.0
docling>=2.0.0
chromadb>=1.0.12
llama-index-core>=0.14.18
llama-index-vector-stores-chroma>=0.5.5
openai>=1.79.0
ruff>=0.5.0
